Added support for Beryl AX

+++ glinet-remove-chinalock.sh
@@ -10,7 +10,7 @@
 # Author: Admon
 # Updated: 2024-05-05
 # Date: 2024-05-05
-SCRIPT_VERSION="2024.05.05.01"
+SCRIPT_VERSION="2024.05.19.01"
 # ^ Update this version number when you make changes to the script
 #
 # Usage: ./glinet-remove-chinalock.sh [--help] [--new-country-code=<COUNTRY_CODE>] [--country-code=<COUNTRY_CODE>]
@@ -26,7 +26,7 @@ YELLOW='\033[0;33m'
 INFO='\033[0m' # No Color
 BACKUP_FILE="factory_partition_$(date +"%Y%m%d_%H%M%S").img"
 
-# FuINFOtions
+# Functions
 invoke_intro() {
     echo "┌────────────────────────────────────────────────────────────────────────┐"
     echo "│ GL.iNet router script by Admon 🦭 for the GL.iNet community            |"
@@ -50,25 +50,30 @@ invoke_help() {
 
 invoke_convert_countrycode() {
     COUNTRY_CODE_HEX=$(echo -n "$1" | hexdump -v -e '/1 "%02X"')
-    # Add FF to beginning of the country code
-    COUNTRY_CODE_HEX="FF${COUNTRY_CODE_HEX}"
     echo $COUNTRY_CODE_HEX
 }
 
 invoke_find_partition() {
-    log "INFO" "Searching for the factory partition ..."
+    log "INFO" "Searching for factory partition by label ..."
+    # Try to find the factory partition by label
     FACTORY_PARTITION=$(blkid -t PARTLABEL="factory" -o device)
     if [ -z "$FACTORY_PARTITION" ]; then
-        log "ERROR" "Could not find the factory partition"
-        log "ERROR" "Please report this issue to the GitHub repository."
-        exit 1
-    else
-        log "SUCCESS" "Found the factory partition at $FACTORY_PARTITION"
+        log "WARNING" "Could not find factory partition by label"
+        log "INFO" "Searching for the factory partition by MTD ..."
+        # Try to find the factory partition by MTD
+        FACTORY_MTD=$(awk '/Factory/ {gsub(":", "", $1); print $1}' /proc/mtd)
+        FACTORY_PARTITION="/dev/mtdblock${FACTORY_MTD:3}"
+        if [ ! -e "$FACTORY_PARTITION" ]; then
+            log "ERROR" "Could not find factory partition"
+            log "ERROR" "Please report this issue to the GitHub repository."
+            exit 1
+        fi
     fi
+    log "SUCCESS" "Found factory partition at $FACTORY_PARTITION"
 }
 
 invoke_backup() {
-    log "INFO" "Backing up the factory partition ..."
+    log "INFO" "Backing up factory partition ..."
     dd if=$FACTORY_PARTITION of=$BACKUP_FILE bs=2M > /dev/null 2>&1
     # Check if the backup was successful by checking the file existeINFOe
     if [ -f $BACKUP_FILE ]; then
@@ -104,29 +109,31 @@ invoke_find_countrycode()
     local DEVICE=$1
     local COUNTRY_CODE=$2
     local COUNTRY_CODE_HEX=$3
-    log "INFO" "Searching for the country code $COUNTRY_CODE ($COUNTRY_CODE_HEX) in the factory partition $DEVICE ..."
+    log "INFO" "Searching for country code $COUNTRY_CODE ($COUNTRY_CODE_HEX) in factory partition $DEVICE ..."
     BINARY_OFFSET=$(hexdump -ve '1/1 "%.2X"' $DEVICE | tr -d '\n' | sed -n "s/\(.*\)$COUNTRY_CODE_HEX\(.*\)/\1/p" | wc -c)
     if [ -z $BINARY_OFFSET ]; then
-        log "ERROR" "Could not find the country code ($COUNTRY_CODE_HEX) in the factory partition"
+        log "WARNING" "Could not find country code ($COUNTRY_CODE_HEX) in factory partition"
         log "ERROR" "Please report this issue to the GitHub repository."
+        log "ERROR" "Error: BINARY_OFFSET is empty"
         exit 1
     else
         # Divide by 2 to get the actual offset
-        BINARY_OFFSET=$((BINARY_OFFSET / 2))
+
         # Offset should never be 0
         if [ $BINARY_OFFSET -eq 0 ]; then
-            log "ERROR" "Could not find the country code ($COUNTRY_CODE_HEX) in the factory partition"
-            log "ERROR" "Please report this issue to the GitHub repository."
-            exit 1
+            log "WARNING" "Could not find country code ($COUNTRY_CODE_HEX) in factory partition"
+        else
+            BINARY_OFFSET=$((BINARY_OFFSET / 2))
+            BINARY_OFFSET=$((BINARY_OFFSET + 1))
+            log "SUCCESS" "Found country code ($COUNTRY_CODE_HEX) at offset $BINARY_OFFSET"
         fi
-        log "SUCCESS" "Found the country code ($COUNTRY_CODE_HEX) at offset $BINARY_OFFSET"
     fi
 }
 
 invoke_ask() {
     echo -e "${RED}W A R N I N G${INFO}"
     echo -e "${RED}This script is HIGLY EXPERIMENTAL and NOT OFFICIALLY SUPPORTED by GL.iNet!${INFO}"
-    echo -e "${RED}It could DESTROY your device and VOID your warranty!${INFO}"
+    echo -e "${RED}It could DESTROY your device and WILL VOID your warranty!${INFO}"
     echo -e "${RED}Even UNBRICKING might not be possible!${INFO}"
     echo -e "${RED}Make sure to copy the backup file to a safe place!${INFO}"
     echo -e "${RED}After running this script you need to set up the device again!${INFO}"
@@ -152,22 +159,37 @@ invoke_done() {
 }
 
 invoke_update() {
-     SCRIPT_VERSION_NEW=$(curl -s "https://raw.githubusercontent.com/Admonstrator/glinet-remove-chinalock/main/glinet-remove-chinalock.sh" | grep -o 'SCRIPT_VERSION="[0-9]\{4\}\.[0-9]\{2\}\.[0-9]\{2\}\.[0-9]\{2\}"' | cut -d '"' -f 2 || echo "Failed to retrieve script version")
-    if [ "$SCRIPT_VERSION_NEW" != "$SCRIPT_VERSION" ]; then
-        log "INFO" "A new version of the script is available: $SCRIPT_VERSION_NEW"
-        log "INFO" "Updating the script ..."
-        wget -qO /tmp/update-tailscale.sh "https://raw.githubusercontent.com/Admonstrator/glinet-remove-chinalock/main/glinet-remove-chinalock.sh"
-        # Get current script path
-        SCRIPT_PATH=$(readlink -f "$0")
-        # Replace current script with updated script
-        rm "$SCRIPT_PATH"
-        mv /tmp/update-tailscale.sh "$SCRIPT_PATH"
-        chmod +x "$SCRIPT_PATH"
-        log "INFO" "The script has been updated. It will now restart ..."
-        sleep 3
-        exec "$SCRIPT_PATH" "$@"
+    log "INFO" "Checking for script updates ..."
+    SCRIPT_VERSION_NEW=$(curl -s "https://raw.githubusercontent.com/Admonstrator/glinet-remove-chinalock/main/glinet-remove-chinalocksh" | grep -o 'SCRIPT_VERSION="[0-9]\{4\}\.[0-9]\{2\}\.[0-9]\{2\}\.[0-9]\{2\}"' | cut -d '"' -f 2 || echo "Failed to retrieve scriptversion")
+    if [ -n "$SCRIPT_VERSION_NEW" ] && [ "$SCRIPT_VERSION_NEW" != "$SCRIPT_VERSION" ]; then
+       log "INFO" "A new version of the script is available: $SCRIPT_VERSION_NEW"
+       log "INFO" "Updating the script ..."
+       wget -qO /tmp/glinet-remove-chinalock.sh "https://raw.githubusercontent.com/Admonstrator/glinet-remove-chinalock/main/glinet-remove-chinalock.sh"
+       # Get current script path
+       SCRIPT_PATH=$(readlink -f "$0")
+       # Replace current script with updated script
+       rm "$SCRIPT_PATH"
+       mv /tmp/glinet-remove-chinalock.sh "$SCRIPT_PATH"
+       chmod +x "$SCRIPT_PATH"
+       log "INFO" "The script has been updated. It will now restart ..."
+       sleep 3
+       exec "$SCRIPT_PATH" "$@"
     else
         log "INFO" "The script is up to date"
+    fi
+}
+
+invoke_check_dependencies(){
+    # Check if the provided country codes are only 2 characters long
+    if [ ${#COUNTRY_CODE} -ne 2 ] || [ ${#COUNTRY_CODE_NEW} -ne 2 ]; then
+        log "ERROR" "Country codes must be 2 characters long"
+        exit 1
+    fi
+
+    # Check if awk is installed
+    if ! command -v awk > /dev/null; then
+        log "ERROR" "awk is not installed"
+        exit 1
     fi
 }
 
@@ -209,7 +231,9 @@ while [[ "$#" -gt 0 ]]; do
     esac
 done
 
+# Main part
 invoke_update
+invoke_check_dependencies
 invoke_intro
 invoke_ask
 log "INFO" "Changing the country code from $COUNTRY_CODE to $COUNTRY_CODE_NEW"
@@ -217,7 +241,21 @@ COUNTRY_CODE_HEX=$(invoke_convert_countrycode $COUNTRY_CODE)
 COUNTRY_CODE_NEW_HEX=$(invoke_convert_countrycode $COUNTRY_CODE_NEW)
 invoke_find_partition
 invoke_backup
-invoke_find_countrycode $FACTORY_PARTITION $COUNTRY_CODE $COUNTRY_CODE_HEX
+
+log "INFO" "Try 1#"
+log "INFO" "Due to different encoding of the country code, the script will search twice"
+invoke_find_countrycode $FACTORY_PARTITION $COUNTRY_CODE FF${COUNTRY_CODE_HEX}
+if [ $BINARY_OFFSET -eq "0" ]; then
+    log "INFO" "Try #2"
+    invoke_find_countrycode $FACTORY_PARTITION $COUNTRY_CODE 00${COUNTRY_CODE_HEX}
+    if [ $BINARY_OFFSET -eq "0" ]; then
+        log "ERROR" "Could not find country code in factory partition"
+        log "ERROR" "Please report this issue to the GitHub repository."
+        exit 1
+    fi
+fi
+
+# All set, let's go!
 invoke_remove_china_lock
 invoke_done
 exit 0
